refactor(player_commands)
- Remover (temporalmente?) la implementación de la interfaz Serializable
  en Command
- Los comandos reciben Player por parámetro. El modelo hace
  cmd.execute(player) para que la ejecución del comando se propague por
  el resto de los componentes del modelo (según corresponda)
