+++ common/player_commands/move.cpp
@@ -1,0 +1,8 @@
+#include "move.h"
+
+#include "../network/serializer.h"
+
+Move::Move(Direction d): dir(d) {}
+
+void Move::execute(Player& p) const { p.move(dir); }
+Direction Move::get_dir() const { return dir; }
