+++ common/position.h
@@ -4,7 +4,7 @@
 struct Position {
     float x;
     float y;
-    Position(const float& x, const float& y);
+    Position(const float& x, const float& y): x(x), y(y) {}
 };
 
 #endif
